feat: tag 가져오는 기능 구현
