+++ frontend/src/hooks/review/index.ts
@@ -1,3 +1,4 @@
 export { default as useReviewFavorite } from './useReviewFavorite';
 export { default as useDisplayTag } from './useDisplayTag';
 export { default as useReviewTags } from './useReviewTags';
+export { default as useSelectedTags } from './useSelectedTags';
